# docs/swagger.yaml
basePath: /
definitions:
  dnsapi.CreateRecordRequest:
    properties:
      name:
        example: www
        type: string
      prio:
        example: 0
        type: integer
      ttl:
        example: 3600
        type: integer
      type:
        example: A
        type: string
      value:
        example: 1.2.3.4
        type: string
    type: object
  dnsapi.CreateZoneRequest:
    properties:
      abuse_email:
        example: admin@example.com
        type: string
      domain:
        example: example.com
        type: string
      owner:
        example: alice
        type: string
      tags:
        example: tag1,tag2
        type: string
    type: object
  dnsapi.DNSCheckRecord:
    properties:
      db_value:
        type: string
      error:
        type: string
      fqdn:
        type: string
      name:
        type: string
      primary_ok:
        type: boolean
      primary_values:
        items:
          type: string
        type: array
      record_id:
        type: integer
      secondary_ok:
        type: boolean
      secondary_values:
        items:
          type: string
        type: array
      supported:
        type: boolean
      type:
        type: string
    type: object
  dnsapi.DNSCheckResponse:
    properties:
      domain:
        type: string
      primary_ns:
        type: string
      records:
        items:
          $ref: '#/definitions/dnsapi.DNSCheckRecord'
        type: array
      secondary_ns:
        items:
          type: string
        type: array
      zone_id:
        type: integer
    type: object
  dnsapi.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  dnsapi.MessageResponse:
    properties:
      message:
        type: string
    type: object
  dnsapi.Record:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      prio:
        type: integer
      ttl:
        type: integer
      type:
        description: A, AAAA, CNAME, TXT, SRV
        type: string
      updated_at:
        type: string
      value:
        type: string
    type: object
  dnsapi.UpdateRecordRequest:
    properties:
      name:
        example: www
        type: string
      prio:
        example: 0
        type: integer
      ttl:
        example: 3600
        type: integer
      value:
        example: 1.2.3.4
        type: string
    type: object
  dnsapi.UpdateZoneRequest:
    properties:
      abuse_email:
        example: admin@example.com
        type: string
      owner:
        example: alice
        type: string
      tags:
        example: tag1,tag2
        type: string
    type: object
  dnsapi.Zone:
    properties:
      abuse_email:
        type: string
      committed_serial:
        type: string
      created_at:
        type: string
      delete:
        type: boolean
      domain:
        type: string
      id:
        type: integer
      owner:
        type: string
      records:
        items:
          $ref: '#/definitions/dnsapi.Record'
        type: array
      serial:
        type: string
      tags:
        description: Tags separated by comma
        type: string
      updated_at:
        type: string
    type: object
  dnsapi.ZoneNSStatus:
    properties:
      configured:
        type: boolean
      error:
        type: string
      ip:
        type: string
    type: object
  dnsapi.ZoneStatusResponse:
    properties:
      all_configured:
        type: boolean
      committed_serial:
        type: string
      db_serial:
        type: string
      domain:
        type: string
      needs_commit:
        type: boolean
      primary:
        $ref: '#/definitions/dnsapi.ZoneNSStatus'
      secondary:
        items:
          $ref: '#/definitions/dnsapi.ZoneNSStatus'
        type: array
      zone_id:
        type: integer
    type: object
info:
  contact: {}
  description: Simple DNS API for managing Bind zones and records.
  title: DNS API
  version: "1.1"
paths:
  /bind/config:
    put:
      consumes:
      - text/plain
      description: Receives a rendered named.conf include file, writes it to disk,
        and reloads bind9.
      operationId: put-bind-config
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dnsapi.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Write bind config (secondary mode)
      tags:
      - secondary
  /bind/refresh/{domain}:
    post:
      description: Forces a zone refresh using the configured refresh command.
      operationId: post-zone-refresh
      parameters:
      - description: Zone domain name
        in: path
        name: domain
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dnsapi.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Refresh zone (secondary mode)
      tags:
      - secondary
  /bind/reload:
    post:
      description: Reloads bind9 via the configured reload command.
      operationId: post-bind-reload
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dnsapi.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Reload bind9 (secondary mode)
      tags:
      - secondary
  /bind/zones/{domain}:
    delete:
      description: Removes a zone file from the bind zone directory.
      operationId: delete-zone-file
      parameters:
      - description: Zone domain name
        in: path
        name: domain
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dnsapi.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete zone file (secondary mode)
      tags:
      - secondary
    put:
      consumes:
      - text/plain
      description: Receives a zone file and writes it to the bind zone directory.
      operationId: put-zone-file
      parameters:
      - description: Zone domain name
        in: path
        name: domain
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dnsapi.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Write zone file (secondary mode)
      tags:
      - secondary
...
  /zones/:
    get:
      description: Return all zones including records.
      operationId: get-zones
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dnsapi.Zone'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List zones
      tags:
      - zones
    post:
      consumes:
      - application/json
      description: Create a new zone.
      operationId: create-zone
      parameters:
      - description: Zone payload
        in: body
        name: zone
        required: true
        schema:
          $ref: '#/definitions/dnsapi.CreateZoneRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dnsapi.Zone'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create zone
      tags:
      - zones
  /zones/{zone_id}:
    delete:
      description: Delete a zone and all its records.
      operationId: delete-zone
      parameters:
      - description: Zone ID
        in: path
        name: zone_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dnsapi.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete zone
      tags:
      - zones
    get:
      description: Return one zone including records.
      operationId: get-zone
      parameters:
      - description: Zone ID
        in: path
        name: zone_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dnsapi.Zone'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get zone
      tags:
      - zones
    put:
      consumes:
      - application/json
      description: Update zone metadata (tags and abuse email only). Use the records
        endpoints to manage records.
      operationId: update-zone
      parameters:
      - description: Zone ID
        in: path
        name: zone_id
        required: true
        type: integer
      - description: Zone metadata to update
        in: body
        name: zone
        required: true
        schema:
          $ref: '#/definitions/dnsapi.UpdateZoneRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dnsapi.Zone'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update zone
      tags:
      - zones
  /zones/{zone_id}/commit:
    put:
      description: Write zone changes into DNS servers.
      operationId: commit-zone
      parameters:
      - description: Zone ID
        in: path
        name: zone_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dnsapi.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Commit zone
      tags:
      - zones
  /zones/{zone_id}/records/:
    get:
      description: Return all records for a zone.
      operationId: get-records
      parameters:
      - description: Zone ID
        in: path
        name: zone_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dnsapi.Record'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List records
      tags:
      - records
    post:
      consumes:
      - application/json
      description: Create a new record in a zone. Call commit afterwards to write
        it to bind.
      operationId: create-record
      parameters:
      - description: Zone ID
        in: path
        name: zone_id
        required: true
        type: integer
      - description: Record payload
        in: body
        name: record
        required: true
        schema:
          $ref: '#/definitions/dnsapi.CreateRecordRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dnsapi.Record'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create record
      tags:
      - records
  /zones/{zone_id}/records/{record_id}:
    delete:
      description: Delete a record within a zone.
      operationId: delete-record
      parameters:
      - description: Zone ID
        in: path
        name: zone_id
        required: true
        type: integer
      - description: Record ID
        in: path
        name: record_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dnsapi.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete record
      tags:
      - records
    get:
      description: Return one record within a zone.
      operationId: get-record
      parameters:
      - description: Zone ID
        in: path
        name: zone_id
        required: true
        type: integer
      - description: Record ID
        in: path
        name: record_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dnsapi.Record'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get record
      tags:
      - records
    put:
      consumes:
      - application/json
      description: Update an existing record within a zone. Type cannot be changed;
        delete and recreate instead. Call commit afterwards.
      operationId: update-record
      parameters:
      - description: Zone ID
        in: path
        name: zone_id
        required: true
        type: integer
      - description: Record ID
        in: path
        name: record_id
        required: true
        type: integer
      - description: Record fields to update
        in: body
        name: record
        required: true
        schema:
          $ref: '#/definitions/dnsapi.UpdateRecordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dnsapi.Record'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update record
      tags:
      - records
  /zones/{zone_id}/status:
    get:
      description: Check whether the zone has uncommitted changes and whether it is
        configured on all nameservers.
      operationId: zone-status
      parameters:
      - description: Zone ID
        in: path
        name: zone_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dnsapi.ZoneStatusResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Zone status
      tags:
      - zones
  /zones/{zone_id}/test:
    get:
      description: Compare zone records in the database against what the primary and
        secondary nameservers currently serve.
      operationId: test-zone
      parameters:
      - description: Zone ID
        in: path
        name: zone_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dnsapi.DNSCheckResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Test zone DNS
      tags:
      - zones
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"